% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{academic_year}
\alias{academic_year}
\title{Bereken academisch jaar uit datum}
\usage{
academic_year(date)
}
\arguments{
\item{date}{Date vector of te converteren waarde}
}
\value{
Integer vector met academisch jaren
}
\description{
Bepaalt het academisch jaar op basis van een datum. Een academisch jaar
begint op 1 september. Datums vanaf september worden toegewezen aan het
huidige jaar, datums vóór september aan het vorige jaar.
}
\examples{
\dontrun{
  academic_year(as.Date("2024-09-01"))  # Returns 2024
  academic_year(as.Date("2024-08-31"))  # Returns 2023
}
}
\keyword{internal}
