% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{audit_enrollments}
\alias{audit_enrollments}
\title{Controleer datakwaliteit van inschrijvingsgegevens}
\usage{
audit_enrollments(enrollments)
}
\arguments{
\item{enrollments}{Data frame met ruwe inschrijvingsgegevens}
}
\value{
Data frame met vertaalde kolom namen
}
\description{
Voert basale kwaliteitscontroles uit op inschrijvingsgegevens, inclusief
controle op lege kolommen en hoog percentage missende waarden.
Audit informatie wordt als messages/warnings getoond.
}
