% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{audit_rio}
\alias{audit_rio}
\title{Controleer en vertaal RIO kolommen}
\usage{
audit_rio(rio_data)
}
\arguments{
\item{rio_data}{Data frame met ruwe RIO data}
}
\value{
Data frame met vertaalde kolom namen
}
\description{
Vertaalt kolom namen van RIO data van DUO export formaat naar intern formaat.
}
\keyword{internal}
