% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_enrollments_calculations}
\alias{combine_enrollments_calculations}
\title{Voeg berekende velden toe}
\usage{
combine_enrollments_calculations(enrollments)
}
\arguments{
\item{enrollments}{Data frame met gecombineerde inschrijvingsgegevens}
}
\value{
Data frame met toegevoegde berekende velden
}
\description{
Berekent afgeleide indicatoren voor study success en switches.
}
