% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_enrollments_final}
\alias{combine_enrollments_final}
\title{Finale transformaties}
\usage{
combine_enrollments_final(enrollments)
}
\arguments{
\item{enrollments}{Data frame met berekende inschrijvingsgegevens}
}
\value{
Data frame gereed voor export
}
\description{
Voert laatste datatype conversies en kolomselectie uit.
}
