% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_enrollments_rio}
\alias{combine_enrollments_rio}
\title{Combineer inschrijvingsgegevens met RIO data}
\usage{
combine_enrollments_rio(enrollments, rio_per_jaar)
}
\arguments{
\item{enrollments}{Data frame met voorbereide inschrijvingsgegevens}

\item{rio_per_jaar}{Data frame met RIO data per academisch jaar}
}
\value{
Data frame met verrijkte inschrijvingsgegevens
}
\description{
Voegt RIO referentiedata toe aan inschrijvingsgegevens via left join.
}
