% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_metadata}
\alias{detect_metadata}
\title{Detecteer metadata uit inschrijvingsdata}
\usage{
detect_metadata(enrollments)
}
\arguments{
\item{enrollments}{Data frame met inschrijvingsgegevens}
}
\value{
Lijst met:
\item{year}{Maximaal inschrijvingsjaar, of NULL als niet gevonden}
\item{institution_brin}{Meest voorkomende BRIN code, of NULL als niet gevonden}
}
\description{
Detecteert de BRIN code van de instelling en het academisch jaar uit
een inschrijvings data frame. Werkt met zowel ruwe 1CHO kolomnamen
(lowercase) als vertaalde kolomnamen (INS_ prefix).
}
