% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_rio}
\alias{get_rio}
\title{Get RIO referentiedata}
\usage{
get_rio(force_download = FALSE)
}
\arguments{
\item{force_download}{Logisch. Forceer download van DUO? Standaard: FALSE}
}
\value{
Data frame met RIO referentiedata
}
\description{
Haal actuele RIO (Register Instellingen en Opleidingen) data op.
Probeert eerst gebundelde data te lezen uit package installatie.
Als niet beschikbaar, download dan van DUO (zonder op te slaan).
}
\examples{
\dontrun{
  # Get RIO data (from package or download)
  rio <- get_rio()

  # Force download from DUO
  rio <- get_rio(force_download = TRUE)
}
}
