% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_enrollments_mapping}
\alias{prepare_enrollments_mapping}
\title{Pas mapping tables toe op inschrijvingsgegevens}
\usage{
prepare_enrollments_mapping(enrollments)
}
\arguments{
\item{enrollments}{Data frame met ruwe inschrijvingsgegevens}
}
\value{
Data frame met gemapte waarden
}
\description{
Vertaalt ruwe veldwaarden naar leesbare labels met behulp van mapping tables.
}
\examples{
\dontrun{
  enrollments <- read.csv2("data.csv")
  enrollments_mapped <- prepare_enrollments_mapping(enrollments)
}
}
