% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_enrollments_supplemental}
\alias{prepare_enrollments_supplemental}
\title{Bereid inschrijvingen voor - supplemental fields}
\usage{
prepare_enrollments_supplemental(enrollments, year, institution_brin)
}
\arguments{
\item{enrollments}{Data frame met gemapte inschrijvingsgegevens}

\item{year}{Academisch jaar voor configuratie}

\item{institution_brin}{BRIN code van de instelling}
}
\value{
Data frame met supplemental fields
}
\description{
Voeg aanvullende berekende velden toe aan inschrijvingsgegevens.
}
