% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_rio}
\alias{prepare_rio}
\title{Bereid RIO referentiedata voor}
\usage{
prepare_rio(
  rio,
  year = 2024,
  institution_brin = "21XX",
  create_synthetic = TRUE
)
}
\arguments{
\item{rio}{Data frame met ruwe RIO data}

\item{year}{Academisch jaar voor maximale datum. Standaard: 2024}

\item{institution_brin}{BRIN code van de instelling. Standaard: "21XX"}

\item{create_synthetic}{Logisch. Maak synthetische test rows? Standaard: TRUE}
}
\value{
Lijst met:
\item{rio}{Voorbereide RIO data}
\item{rio_per_jaar}{RIO data per academisch jaar}
}
\description{
Transformeert ruwe RIO (Register Instellingen en Opleidingen) data naar
bruikbaar formaat met datumvelden, status codes, en opleidingsnamen met
niveau prefix (B/M/AD).
}
\keyword{internal}
