% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_values}
\alias{round_values}
\title{Rond waarden af (negatief naar beneden, positief naar boven)}
\usage{
round_values(x)
}
\description{
Rond waarden af (negatief naar beneden, positief naar boven)
}
\keyword{internal}
