% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Start de prep1cho Shiny app}
\usage{
run_app(...)
}
\arguments{
\item{...}{Extra argumenten doorgegeven aan \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
Onzichtbaar. Wordt aangeroepen voor het starten van de app.
}
\description{
Interactieve data verkenner voor prep1cho pipeline output.
Upload een CSV of FST bestand, of gebruik automatisch de synthetische data.
}
