% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_colnames_documentation}
\alias{translate_colnames_documentation}
\title{Vertaal kolom namen met documentatie bestand}
\usage{
translate_colnames_documentation(
  data,
  doc_file,
  doc_dir = "metadata/assertions"
)
}
\arguments{
\item{data}{Data frame met kolommen in export formaat}

\item{doc_file}{Bestandsnaam van documentatie CSV (bijv. "Documentatie_RIO.csv")}

\item{doc_dir}{Directory waar documentatie bestand staat. Standaard: "metadata/assertions"}
}
\value{
Data frame met vertaalde kolom namen
}
\description{
Algemene functie om kolom namen te vertalen van export formaat naar
intern formaat, gebruik makend van een documentatie CSV bestand.
}
\keyword{internal}
