% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{translate_enrollments_colnames}
\alias{translate_enrollments_colnames}
\title{Translate enrollment column names from 1CHO to internal format}
\usage{
translate_enrollments_colnames(enrollments)
}
\arguments{
\item{enrollments}{Data frame met 1CHO enrollment data (export kolom namen)}
}
\value{
Data frame met vertaalde kolom namen (internal format)
}
\description{
Translate enrollment column names from 1CHO to internal format
}
\keyword{internal}
