% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{translate_rio_colnames}
\alias{translate_rio_colnames}
\title{Translate RIO column names from new to old format}
\usage{
translate_rio_colnames(rio_data)
}
\arguments{
\item{rio_data}{Data frame with RIO data (new column names)}
}
\value{
Data frame with translated column names (old format)
}
\description{
Translate RIO column names from new to old format
}
\keyword{internal}
